Added prep quests
Fixed the Run to Org step so it actually works as expected. Added a note that will show only if the DF staring quest isn't available, meaning the expansion isn't out yet, to tell the user to get and complete the 3 quests in Org. Added steps to hand in the 3 quests if the user has them in their log, can't check for completion.