Update Waking Shores guide.
Added a waypoint for arriving in the Dragon Isles.
Added a waypoint to guide the user around the cliff to Pathfinder Poppy
Updated the first waypoint for Baron Crustcore to be less miss-able.
Moved the waypoint for the My First Real Emergency RP.
Added the quest item to Veteran Reinforcements.
Added the quest item and a way point to Fighting Fire with... Water.
Added in Dragon Riding Glyphs to the guide.
Added smart optional steps for when you're meant to get on a dragon riding mount.
Added the step to collect the Rostrum if you don't have it already.
Updated the icon for the talk with Celormu step in Dragon Glyphs and You
Updated the waypoint for the ferry step of Return to the Ruby Lifeshrine.
Added a step to collect the tower Dragon Glyph.
Changed the first step in Training Wings to use a vehicle icon.
Added a step to collect the Flashfrost Enclave glyph.
Changed the icon for The Primary Threat.
Potentially added a working target button for Proto-fight.
Added quest item to Cut Off the Head.
Added a step to collect the Obsidian Bulwark glyph.
Added a note about the RP and Wrathion quest bug to the Wrathion Awaits since it doesn't seem like Blizz is going to fix this.
Fix the Talon Strike quest not being unpinned before the step to turn it in.
Updated the first waypoint for The Courage of One's Convictions so that it should actually trigger when the quest step updates.
Updated icons for Taking the Walls steps.
Removed the waypoint and instructions to fly with Wrathion during Taking the Walls since you don't have to and it's faster to use dragon riding.
Updated icons for Obsidian Oathstone.
Removed some waypoints for flying to Wrathion during Obsidian Oathstone since it doesn't matter which way you fly to him.
Added a note about the Wrathion bug when handing in Obsidian Oathstone.
Updated icons for A Shattered Past.
Updated icons for The Spark.
Added a step in The Spark for the RP.
Updated icons for Restoring the Faith.
Updated icons for Heir Apparent.
Updated icons for Claimant to the Throne.
Updated icons for Black Wagon Flight.
Added a step for the Rubyscale Outpost dragon glyph.
Added a step for the Mirewood Fen dragon glyph.