Update the Ohn'ahran Plains guide.
Split Supplies for the Journey into individual steps instead of using a CC.
Updated icons for Making Introductions.
Updated icons for Ora-cull.
Updated icon for A Centaur's Best Friend.
Updated icons for Release the Hounds.
Split Release the Hounds into individual steps.
Updated notes for accepting and handing in The Sole Mender and Return to Roscha.
Updated the icon for Return to Roscha.
Added a note about vigor to the turn in before By Broken Road.
Updated icon for By Broken Road.
Updated icon and hand in note for Omens on the Wind.
Updated the icon for Maruuaki.
Updated the icons for Omens and Incense
Added a step for Omens and Incense to track looting the blooms.
Added a step for Omens and Incense to use the blooms as a quest item.
Added a vigor note to the Emberwatch turn in.
Updated the note and icon for Tools of the Tirade.
Updated note for The Relic Inquiry.
Moved the set hearthstone step to a slightly better spot.
Updated icons for Honoring Our Ancestors.
Updated icons and step names for Clans of the Plains.
Updated icon for The Emissary's Arrival.
Updated icon for The Khanam Matra.
Updated icon for Return to Mender.
Updated note and added quest item for Definitely Eternal Slumber.
Updated note for And Stay Dead!
Updated icons for The Calm Before the Storm.
Updated the icons for Rellen, the Learned.
Updated the icons for Pessimistic Mystic.
Updated icons and notes for Mystic Mystery.
Updated note for Taken By Storm.
Updated icons for Fowl Sorcery.
Split Fowl Sorcery into individual steps.
Updated the Prozela Galeshot quest to track correctly.
Added a step to hearthstone back to Maruukai after Oh No, Ohn'ahra.
Updated icons for Chasing the Wind.
Updated icon for Blowing of the Horn.
Updated icons for Green Dragon Down.
Updated icons for Counting Sheep.
Added target and quest item buttons for Counting Sheep.
Updated icons for Last Resort Analysis.
Added a set hearthstone step in Pinewood Post.
Updated icons for Show of Storm.
Updated icon for With the Wind At Our Backs.
Removed the Shady Sanctuary to speak with Aronus.
Corrected ordering and added a note for The Primalist Front.
Update icon and added a note for Stormbreaker.